+++ Source/Core/EventDispatcher.cpp
@@ -27,10 +27,10 @@
 
 #include "precompiled.h"
 #include "EventDispatcher.h"
-#include <Rocket/Core/Element.h>
-#include <Rocket/Core/Event.h>
-#include <Rocket/Core/EventListener.h>
-#include <Rocket/Core/Factory.h>
+#include "../../Include/Rocket/Core/Element.h"
+#include "../../Include/Rocket/Core/Event.h"
+#include "../../Include/Rocket/Core/EventListener.h"
+#include "../../Include/Rocket/Core/Factory.h"
 
 namespace Rocket {
 namespace Core {
@@ -131,7 +131,7 @@ bool EventDispatcher::DispatchEvent(Element* target_element, const String& name,
 	event->SetPhase(Event::PHASE_CAPTURE);
 	// Capture phase - root, to target (only events that have registered as capture events)
 	// Note: We walk elements in REVERSE as they're placed in the list from the elements parent to the root
-	for (int i = elements.size() - 1; i >= 0 && event->IsPropagating(); i--) 
+	for (int i = (int)elements.size() - 1; i >= 0 && event->IsPropagating(); i--) 
 	{
 		EventDispatcher* dispatcher = elements[i]->GetEventDispatcher();
 		event->SetCurrentElement(elements[i]);
